+++ visualization/app/codeCharta/ui/legendPanel/legendPanel.component.ts
@@ -22,6 +22,10 @@ import {
 	IsAttributeSideBarVisibleSubscriber
 } from "../../state/store/appSettings/isAttributeSideBarVisible/isAttributeSideBarVisible.service"
 import { isDeltaState } from "../../model/files/files.helper"
+import {
+	ColorMetricService,
+	ColorMetricSubscriber
+} from "../../state/store/dynamicSettings/colorMetric/colorMetric.service";
 
 export interface PackageList {
 	colorPixel: string
@@ -31,6 +35,7 @@ export interface PackageList {
 export class LegendPanelController
 	implements
 		IsAttributeSideBarVisibleSubscriber,
+		ColorMetricSubscriber,
 		ColorRangeSubscriber,
 		InvertColorRangeSubscriber,
 		MarkedPackagesSubscriber,
@@ -40,6 +45,7 @@ export class LegendPanelController
 		isLegendVisible: boolean
 		isSideBarVisible: boolean
 		isDeltaState: boolean
+		colorMetric: string
 		colorRange: ColorRange
 		invertColorRange: boolean
 		packageLists: PackageList[]
@@ -48,6 +54,7 @@ export class LegendPanelController
 		isLegendVisible: false,
 		isSideBarVisible: null,
 		isDeltaState: null,
+		colorMetric: null,
 		colorRange: null,
 		invertColorRange: null,
 		packageLists: null,
@@ -56,12 +63,17 @@ export class LegendPanelController
 
 	/* @ngInject */
 	constructor(private $rootScope: IRootScopeService, private storeService: StoreService) {
+		ColorMetricService.subscribe(this.$rootScope, this)
 		ColorRangeService.subscribe(this.$rootScope, this)
 		InvertColorRangeService.subscribe(this.$rootScope, this)
 		IsAttributeSideBarVisibleService.subscribe(this.$rootScope, this)
 		MarkedPackagesService.subscribe(this.$rootScope, this)
 		WhiteColorBuildingsService.subscribe(this.$rootScope, this)
 		InvertDeltaColorsService.subscribe(this.$rootScope, this)
+	}
+
+	onColorMetricChanged(colorMetric: string) {
+		this._viewModel.colorMetric = colorMetric
 	}
 
 	onColorRangeChanged(colorRange: ColorRange) {